.ipynb_checkpoints/requirements-checkpoint.txt
numpy
Pillow
git+https://github.com/ildoonet/pytorch-randaugment
timm
torch
torcheval
torchvision
catalyst
transformers
torchmetrics
openpyxl
pandas